# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ActionOutcome": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message",
          "backend",
          "action_id"
        ],
        "title": "ActionOutcome",
        "type": "object"
      },
      "ActionRequest": {
        "properties": {
          "action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Id"
          },
          "action_type": {
            "enum": [
              "create_ticket",
              "send_reply",
              "escalate_to_human"
            ],
            "title": "Action Type",
            "type": "string"
          },
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          },
          "requires_confirmation": {
            "default": true,
            "title": "Requires Confirmation",
            "type": "boolean"
          }
        },
        "required": [
          "action_type"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
      "ConfirmationRequest": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "action_hash": {
            "title": "Action Hash",
            "type": "string"
          },
          "decision": {
            "enum": [
              "approve",
              "reject",
              "modify"
            ],
            "title": "Decision",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "decision",
          "version",
          "action_hash"
        ],
        "title": "ConfirmationRequest",
        "type": "object"
      },
      "ExecutionState": {
        "enum": [
          "queued",
          "planning",
          "executing",
          "verifying",
          "gating",
          "web_verifying",
          "no_info_check",
          "pending_confirmation",
          "action_executing",
          "completed",
          "escalated",
          "cancelled",
          "failed"
        ],
        "title": "ExecutionState",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PendingConfirmation": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionRequest"
          },
          "action_hash": {
            "title": "Action Hash",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "action",
          "action_hash",
          "expires_at"
        ],
        "title": "PendingConfirmation",
        "type": "object"
      },
      "RunRecord": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "pending_confirmation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingConfirmation"
              },
              {
                "type": "null"
              }
            ]
          },
          "request": {
            "$ref": "#/components/schemas/RunRequest"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SupportResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ExecutionState",
            "default": "queued"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "request"
        ],
        "title": "RunRecord",
        "type": "object"
      },
      "RunRequest": {
        "properties": {
          "do_action": {
            "default": true,
            "title": "Do Action",
            "type": "boolean"
          },
          "dry_run": {
            "default": true,
            "title": "Dry Run",
            "type": "boolean"
          },
          "identity": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identity"
          },
          "query": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "use_web": {
            "default": true,
            "title": "Use Web",
            "type": "boolean"
          },
          "vertical": {
            "anyOf": [
              {
                "enum": [
                  "gov",
                  "saas",
                  "ec"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vertical"
          }
        },
        "required": [
          "query"
        ],
        "title": "RunRequest",
        "type": "object"
      },
      "SupportResult": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "action_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "action_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Result"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "title": "Citations",
            "type": "array"
          },
          "contradiction": {
            "default": false,
            "title": "Contradiction",
            "type": "boolean"
          },
          "decision": {
            "default": "escalate",
            "enum": [
              "answer",
              "escalate"
            ],
            "title": "Decision",
            "type": "string"
          },
          "forced_escalate": {
            "default": false,
            "title": "Forced Escalate",
            "type": "boolean"
          },
          "groundedness": {
            "default": 0.0,
            "title": "Groundedness",
            "type": "number"
          },
          "groundedness_decided": {
            "default": 0,
            "title": "Groundedness Decided",
            "type": "integer"
          },
          "identity_checked": {
            "default": false,
            "title": "Identity Checked",
            "type": "boolean"
          },
          "intent": {
            "anyOf": [
              {
                "enum": [
                  "question",
                  "request",
                  "incident"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "no_info_detected": {
            "default": false,
            "title": "No Info Detected",
            "type": "boolean"
          },
          "overall_confidence": {
            "default": 0.0,
            "title": "Overall Confidence",
            "type": "number"
          },
          "source_agreement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Agreement"
          },
          "used_web": {
            "default": false,
            "title": "Used Web",
            "type": "boolean"
          },
          "vertical": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vertical"
          },
          "warning": {
            "default": false,
            "title": "Warning",
            "type": "boolean"
          },
          "web_reused": {
            "default": false,
            "title": "Web Reused",
            "type": "boolean"
          }
        },
        "title": "SupportResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "GRACE Support API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent-support/runs": {
      "get": {
        "operationId": "list_runs_api_agent_support_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunRecord"
                  },
                  "title": "Response List Runs Api Agent Support Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Runs",
        "tags": [
          "agent-support"
        ]
      },
      "post": {
        "operationId": "create_run_api_agent_support_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "agent-support"
        ]
      }
    },
    "/api/agent-support/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_agent_support_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "agent-support"
        ]
      }
    },
    "/api/agent-support/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_agent_support_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "agent-support"
        ]
      }
    },
    "/api/agent-support/runs/{run_id}/confirmations": {
      "post": {
        "operationId": "resolve_confirmation_api_agent_support_runs__run_id__confirmations_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Confirmation",
        "tags": [
          "agent-support"
        ]
      }
    },
    "/api/agent-support/runs/{run_id}/events": {
      "get": {
        "operationId": "stream_events_api_agent_support_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Last-Event-Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events",
        "tags": [
          "agent-support"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/ready": {
      "get": {
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ready Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready"
      }
    }
  }
}